GUI-Testing/requirements-improc.txt
Pillow
pyparsing
